requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
langchain-core==0.3.51
langchain-community==0.3.21
langchain-openai==0.3.12
langchain-google-genai==2.1.2
python-dotenv==1.0.0
pymysql==1.1.0
